WIP: Created mock pigpio to write test data onto the pigpio interface.
I needed some tools for visualizing I2C traffic and making it consistent and slow.
Wanted to make a faster feedback loop that was all in python.
Got starts and stops working.
Got addresses working.
There is a bug, a 1 is getting inserted into the data MSB. Maybe an ack?